Move aggregation validation to its own module
